+++ android/src/main/java/com/doublesymmetry/trackplayer/service/MusicService.kt
@@ -13,7 +13,6 @@ import android.support.v4.media.RatingCompat
 import androidx.annotation.MainThread
 import androidx.core.app.NotificationCompat
 import androidx.core.app.NotificationCompat.PRIORITY_LOW
-import androidx.localbroadcastmanager.content.LocalBroadcastManager
 import com.doublesymmetry.kotlinaudio.models.*
 import com.doublesymmetry.kotlinaudio.models.NotificationButton.*
 import com.doublesymmetry.kotlinaudio.players.QueuedAudioPlayer
@@ -22,16 +21,18 @@ import com.doublesymmetry.trackplayer.extensions.NumberExt.Companion.toMilliseco
 import com.doublesymmetry.trackplayer.extensions.NumberExt.Companion.toSeconds
 import com.doublesymmetry.trackplayer.extensions.asLibState
 import com.doublesymmetry.trackplayer.extensions.find
+import com.doublesymmetry.trackplayer.model.MetadataAdapter
+import com.doublesymmetry.trackplayer.model.PlaybackMetadata
 import com.doublesymmetry.trackplayer.model.Track
 import com.doublesymmetry.trackplayer.model.TrackAudioItem
 import com.doublesymmetry.trackplayer.module.MusicEvents
-import com.doublesymmetry.trackplayer.module.MusicEvents.Companion.EVENT_INTENT
-import com.doublesymmetry.trackplayer.utils.AppForegroundTracker
+import com.doublesymmetry.trackplayer.module.MusicEvents.Companion.METADATA_PAYLOAD_KEY
 import com.doublesymmetry.trackplayer.utils.BundleUtils
 import com.doublesymmetry.trackplayer.utils.BundleUtils.setRating
 import com.facebook.react.HeadlessJsTaskService
 import com.facebook.react.bridge.Arguments
 import com.facebook.react.jstasks.HeadlessJsTaskConfig
+import com.facebook.react.modules.core.DeviceEventManagerModule
 import com.google.android.exoplayer2.ui.R as ExoPlayerR
 import kotlinx.coroutines.*
 import kotlinx.coroutines.flow.flow
@@ -58,6 +59,7 @@ class MusicService : HeadlessJsTaskService() {
     }
 
     private var appKilledPlaybackBehavior = AppKilledPlaybackBehavior.CONTINUE_PLAYBACK
+    private var stopForegroundGracePeriod: Int = DEFAULT_STOP_FOREGROUND_GRACE_PERIOD
 
     var randomSeed: List<Int> = emptyList();
 
@@ -106,20 +108,18 @@ class MusicService : HeadlessJsTaskService() {
      */
     private fun startAndStopEmptyNotificationToAvoidANR() {
         val notificationManager = this.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
-        var name = ""
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            name = "temporary_channel"
             notificationManager.createNotificationChannel(
-                NotificationChannel(name, name, NotificationManager.IMPORTANCE_LOW)
+                NotificationChannel(getString(TrackPlayerR.string.rntp_temporary_channel_id), getString(TrackPlayerR.string.rntp_temporary_channel_name), NotificationManager.IMPORTANCE_LOW)
             )
         }
 
-        val notificationBuilder = NotificationCompat.Builder(this, name)
+        val notificationBuilder = NotificationCompat.Builder(this, getString(TrackPlayerR.string.rntp_temporary_channel_id))
             .setPriority(PRIORITY_LOW)
             .setCategory(Notification.CATEGORY_SERVICE)
             .setSmallIcon(ExoPlayerR.drawable.exo_notification_small_icon)
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
-           notificationBuilder.setForegroundServiceBehavior(NotificationCompat.FOREGROUND_SERVICE_IMMEDIATE)
+            notificationBuilder.foregroundServiceBehavior = NotificationCompat.FOREGROUND_SERVICE_IMMEDIATE
         }
         val notification = notificationBuilder.build()
         startForeground(EMPTY_NOTIFICATION_ID, notification)
@@ -171,7 +171,9 @@ class MusicService : HeadlessJsTaskService() {
 
         appKilledPlaybackBehavior = AppKilledPlaybackBehavior::string.find(androidOptions?.getString(APP_KILLED_PLAYBACK_BEHAVIOR_KEY)) ?: AppKilledPlaybackBehavior.CONTINUE_PLAYBACK
 
-        //TODO: This handles a deprecated flag. Should be removed soon.
+        BundleUtils.getIntOrNull(androidOptions, STOP_FOREGROUND_GRACE_PERIOD_KEY)?.let { stopForegroundGracePeriod = it }
+
+        // TODO: This handles a deprecated flag. Should be removed soon.
         options.getBoolean(STOPPING_APP_PAUSES_PLAYBACK_KEY).let {
             stoppingAppPausesPlayback = options.getBoolean(STOPPING_APP_PAUSES_PLAYBACK_KEY)
             if (stoppingAppPausesPlayback) {
@@ -239,23 +241,23 @@ class MusicService : HeadlessJsTaskService() {
 
         // setup progress update events if configured
         progressUpdateJob?.cancel()
-        val updateInterval = BundleUtils.getIntOrNull(options, PROGRESS_UPDATE_EVENT_INTERVAL_KEY)
+        val updateInterval = BundleUtils.getDoubleOrNull(options, PROGRESS_UPDATE_EVENT_INTERVAL_KEY)
         if (updateInterval != null && updateInterval > 0) {
             progressUpdateJob = scope.launch {
-                progressUpdateEventFlow(updateInterval.toLong()).collect { emit(MusicEvents.PLAYBACK_PROGRESS_UPDATED, it) }
+                progressUpdateEventFlow(updateInterval).collect { emit(MusicEvents.PLAYBACK_PROGRESS_UPDATED, it) }
             }
         }
     }
 
     @MainThread
-    private fun progressUpdateEventFlow(interval: Long) = flow {
+    private fun progressUpdateEventFlow(interval: Double) = flow {
         while (true) {
             if (player.isPlaying) {
                 val bundle = progressUpdateEvent()
                 emit(bundle)
             }
 
-            delay(interval * 1000)
+            delay((interval * 1000).toLong())
         }
     }
 
@@ -435,6 +437,11 @@ class MusicService : HeadlessJsTaskService() {
     }
 
     @MainThread
+    fun updateNowPlayingMetadata(track: Track) {
+        player.notificationManager.overrideMetadata(track.toAudioItem())
+    }
+
+    @MainThread
     fun clearNotificationMetadata() {
         player.notificationManager.hideNotification()
     }
@@ -444,7 +451,7 @@ class MusicService : HeadlessJsTaskService() {
         previousIndex: Int?,
         oldPosition: Double
     ) {
-        var a = Bundle()
+        val a = Bundle()
         a.putDouble(POSITION_KEY, oldPosition)
         if (index != null) {
             a.putInt(NEXT_TRACK_KEY, index)
@@ -456,9 +463,9 @@ class MusicService : HeadlessJsTaskService() {
 
         emit(MusicEvents.PLAYBACK_TRACK_CHANGED, a)
 
-        var b = Bundle()
+        val b = Bundle()
         b.putDouble("lastPosition", oldPosition)
-        if (tracks.size > 0) {
+        if (tracks.isNotEmpty()) {
             b.putInt("index", player.currentIndex)
             b.putBundle("track", tracks[player.currentIndex].originalItem)
             if (previousIndex != null) {
@@ -561,6 +568,10 @@ class MusicService : HeadlessJsTaskService() {
             }
         }
 
+        fun shouldStopForeground(): Boolean {
+            return stopForegroundWhenNotOngoing && (removeNotificationWhenNotOngoing || isForegroundService())
+        }
+
         scope.launch {
             event.notificationStateChange.collect {
                 when (it) {
@@ -572,11 +583,19 @@ class MusicService : HeadlessJsTaskService() {
                             if (player.playWhenReady) {
                                 startForegroundIfNecessary()
                             }
-                        } else if (stopForegroundWhenNotOngoing) {
-                            if (removeNotificationWhenNotOngoing || isForegroundService()) {
-                                @Suppress("DEPRECATION")
-                                stopForeground(removeNotificationWhenNotOngoing)
-                                Timber.d("stopped foregrounding%s", if (removeNotificationWhenNotOngoing) " and removed notification" else "")
+                        } else if (shouldStopForeground()) {
+                            // Allow the application a grace period to complete any actions
+                            // that may necessitate keeping the service in a foreground state.
+                            // For instance, queuing new media (e.g., related music) after the
+                            // user's queue is complete. This prevents the service from potentially
+                            // being immediately destroyed once the player finishes playing media.
+                            scope.launch {
+                                delay(stopForegroundGracePeriod.toLong() * 1000)
+                                if (shouldStopForeground()) {
+                                    @Suppress("DEPRECATION")
+                                    stopForeground(removeNotificationWhenNotOngoing)
+                                    Timber.d("Notification has been stopped")
+                                }
                             }
                         }
                     }
@@ -594,7 +613,6 @@ class MusicService : HeadlessJsTaskService() {
 
                 if (it == AudioPlayerState.ENDED && player.nextItem == null) {
                     emitQueueEndedEvent()
-                    emitPlaybackTrackChangedEvents(null, player.currentIndex, player.position.toSeconds())
                 }
             }
         }
@@ -660,17 +678,41 @@ class MusicService : HeadlessJsTaskService() {
         }
 
         scope.launch {
-            event.onPlaybackMetadata.collect {
-                Bundle().apply {
-                    putString("source", it.source)
-                    putString("title", it.title)
-                    putString("url", it.url)
-                    putString("artist", it.artist)
-                    putString("album", it.album)
-                    putString("date", it.date)
-                    putString("genre", it.genre)
-                    emit(MusicEvents.PLAYBACK_METADATA, this)
+            event.onTimedMetadata.collect {
+                val data = MetadataAdapter.fromMetadata(it)
+                val bundle = Bundle().apply {
+                    putParcelableArrayList(METADATA_PAYLOAD_KEY, ArrayList(data))
                 }
+                emit(MusicEvents.METADATA_TIMED_RECEIVED, bundle)
+
+                // TODO: Handle the different types of metadata and publish to new events
+                val metadata = PlaybackMetadata.fromId3Metadata(it)
+                    ?: PlaybackMetadata.fromIcy(it)
+                    ?: PlaybackMetadata.fromVorbisComment(it)
+                    ?: PlaybackMetadata.fromQuickTime(it)
+
+                if (metadata != null) {
+                    Bundle().apply {
+                        putString("source", metadata.source)
+                        putString("title", metadata.title)
+                        putString("url", metadata.url)
+                        putString("artist", metadata.artist)
+                        putString("album", metadata.album)
+                        putString("date", metadata.date)
+                        putString("genre", metadata.genre)
+                        emit(MusicEvents.PLAYBACK_METADATA, this)
+                    }
+                }
+            }
+        }
+
+        scope.launch {
+            event.onCommonMetadata.collect {
+                val data = MetadataAdapter.fromMediaMetadata(it)
+                val bundle = Bundle().apply {
+                    putBundle(METADATA_PAYLOAD_KEY, data)
+                }
+                emit(MusicEvents.METADATA_COMMON_RECEIVED, bundle)
             }
         }
 
@@ -703,11 +745,20 @@ class MusicService : HeadlessJsTaskService() {
     }
 
     @MainThread
-    private fun emit(event: String?, data: Bundle? = null) {
-        val intent = Intent(EVENT_INTENT)
-        intent.putExtra(EVENT_KEY, event)
-        if (data != null) intent.putExtra(DATA_KEY, data)
-        LocalBroadcastManager.getInstance(this).sendBroadcast(intent)
+    private fun emit(event: String, data: Bundle? = null) {
+        reactNativeHost.reactInstanceManager.currentReactContext
+            ?.getJSModule(DeviceEventManagerModule.RCTDeviceEventEmitter::class.java)
+            ?.emit(event, data?.let { Arguments.fromBundle(it) })
+    }
+
+    @MainThread
+    private fun emitList(event: String, data: List<Bundle> = emptyList()) {
+        val payload = Arguments.createArray()
+        data.forEach { payload.pushMap(Arguments.fromBundle(it)) }
+
+        reactNativeHost.reactInstanceManager.currentReactContext
+            ?.getJSModule(DeviceEventManagerModule.RCTDeviceEventEmitter::class.java)
+            ?.emit(event, payload)
     }
 
     override fun getTaskConfig(intent: Intent?): HeadlessJsTaskConfig {
@@ -775,7 +826,7 @@ class MusicService : HeadlessJsTaskService() {
         const val NEXT_TRACK_KEY = "nextTrack"
         const val POSITION_KEY = "position"
         const val DURATION_KEY = "duration"
-        const val BUFFERED_POSITION_KEY = "buffer"
+        const val BUFFERED_POSITION_KEY = "buffered"
 
         const val TASK_KEY = "TrackPlayer"
 
@@ -794,6 +845,7 @@ class MusicService : HeadlessJsTaskService() {
 
         const val STOPPING_APP_PAUSES_PLAYBACK_KEY = "stoppingAppPausesPlayback"
         const val APP_KILLED_PLAYBACK_BEHAVIOR_KEY = "appKilledPlaybackBehavior"
+        const val STOP_FOREGROUND_GRACE_PERIOD_KEY = "stopForegroundGracePeriod"
         const val PAUSE_ON_INTERRUPTION_KEY = "alwaysPauseOnInterruption"
         const val AUTO_UPDATE_METADATA = "autoUpdateMetadata"
         const val AUTO_HANDLE_INTERRUPTIONS = "autoHandleInterruptions"
@@ -802,5 +854,6 @@ class MusicService : HeadlessJsTaskService() {
         const val IS_PAUSED_KEY = "paused"
 
         const val DEFAULT_JUMP_INTERVAL = 15.0
+        const val DEFAULT_STOP_FOREGROUND_GRACE_PERIOD = 5
     }
 }