+++ example/src/services/PlaybackService.ts
@@ -56,6 +56,26 @@ export async function PlaybackService() {
     console.log('Event.PlaybackState', event);
   });
 
+  TrackPlayer.addEventListener(Event.PlaybackMetadataReceived, (event) => {
+    console.log('[Deprecated] Event.PlaybackMetadataReceived', event);
+  });
+
+  TrackPlayer.addEventListener(Event.MetadataChapterReceived, (event) => {
+    console.log('Event.MetadataChapterReceived', event);
+  });
+
+  TrackPlayer.addEventListener(Event.MetadataTimedReceived, (event) => {
+    console.log('Event.MetadataTimedReceived', event);
+  });
+
+  TrackPlayer.addEventListener(Event.MetadataCommonReceived, (event) => {
+    console.log('Event.MetadataCommonReceived', event);
+  });
+
+  TrackPlayer.addEventListener(Event.PlaybackProgressUpdated, (event) => {
+    console.log('Event.PlaybackProgressUpdated', event);
+  });
+
   TrackPlayer.addEventListener(
     Event.PlaybackMetadataReceived,
     async ({ title, artist }) => {