+++ src/trackPlayer.ts
@@ -224,6 +224,7 @@ export async function updateOptions({
   ...options
 }: UpdateOptions = {}): Promise<void> {
   return TrackPlayer.updateOptions({
+    ...options,
     android: {
       // Handle deprecated alwaysPauseOnInterruption option:
       alwaysPauseOnInterruption:
@@ -238,7 +239,6 @@ export async function updateOptions({
     nextIcon: resolveImportedAsset(options.nextIcon),
     rewindIcon: resolveImportedAsset(options.rewindIcon),
     forwardIcon: resolveImportedAsset(options.forwardIcon),
-    ...options,
   });
 }
 
@@ -269,8 +269,8 @@ export function clearNowPlayingMetadata(): Promise<void> {
 }
 
 /**
- * @deprecated Nominated for removal in the next major version.
- * Use `updateMetadataForTrack` with the current index instead.
+ * Updates the metadata content of the notification (Android) and the Now Playing Center (iOS)
+ * without affecting the data stored for the current track.
  */
 export function updateNowPlayingMetadata(
   metadata: NowPlayingMetadata